requirements.txt
python-dotenv==1.0.1
google-cloud-bigquery==3.27.0
google-cloud-storage==2.19.0
google-cloud-aiplatform==1.74.0
google-generativeai==0.8.3
jupyter==1.1.1
nbconvert==7.16.4
pillow==11.0.0
google-genai==0.3.0
langchain-google-genai==2.0.7
langchain-google-vertexai==2.0.9
Markdown==3.7
langchain==0.3.4
langchain-experimental==0.3.2
langchain-openai==0.2.3
openai==1.52.0
fitz==0.0.1.dev2
PyMuPDF==1.24.13
markdown2==2.5.1
pdfkit==1.0.0
gradio==5.6.0